+++ fews_3di/simulation.py
@@ -107,16 +107,28 @@ class ThreediSimulation:
             self._add_laterals(laterals)
         else:
             logger.info("No lateral timeseries found at %s, skipping.", laterals_csv)
-
-        saved_state_id_file = (
-            self.settings.base_dir / "states" / SAVED_STATE_ID_FILENAME
-        )
-        cold_state_id_file = self.settings.base_dir / "states" / COLD_STATE_ID_FILENAME
-
-        if self.settings.initial_waterlevel:
-            self._add_initial_waterlevel_raster(model_id)
+        if self.settings.api_host == "https://api.staging.3di.live":
+            logger.info("Using the states files in the states_staging folder")
+            saved_state_id_file = (
+                self.settings.base_dir / "states_staging" / SAVED_STATE_ID_FILENAME
+            )
+            cold_state_id_file = (
+                self.settings.base_dir / "states_staging" / COLD_STATE_ID_FILENAME
+            )
         else:
-            logger.info("No initial waterlevel raster predefined")
+            logger.info("Using the states files in the states folder")
+            saved_state_id_file = (
+                self.settings.base_dir / "states" / SAVED_STATE_ID_FILENAME
+            )
+            cold_state_id_file = (
+                self.settings.base_dir / "states" / COLD_STATE_ID_FILENAME
+            )
+
+        # self.add_initial_1d_waterlevels(model_id)
+        # if self.settings.initial_waterlevel:
+        # self._add_initial_waterlevel_raster(model_id)
+        # else:
+        # logger.info("No initial waterlevel raster predefined")
 
         if self.settings.save_state:
             if self.settings.use_last_available_state:
@@ -201,15 +213,19 @@ class ThreediSimulation:
 
     def _create_simulation(self, model_id: int) -> Tuple[int, str]:
         """Return id and url of created simulation."""
+        simulation_template = self.api.simulation_templates_list(
+            simulation__threedimodel__id=model_id
+        ).results[0]
         data = {}
         data["name"] = self.settings.simulationname
+        data["template"] = str(simulation_template.id)
         data["threedimodel"] = str(model_id)
         data["organisation"] = self.settings.organisation
         data["start_datetime"] = self.settings.start.isoformat()
         data["duration"] = str(self.settings.duration)
+        # data["clone_events"] = True
         logger.debug("Creating simulation with these settings: %s", data)
-
-        simulation = self.api.simulations_create(data)
+        simulation = self.api.simulations_from_template(data)
         logger.info("Simulation %s has been created", simulation.url)
         return simulation.id, simulation.url
 
@@ -347,21 +363,28 @@ class ThreediSimulation:
         logger.info("Saved state will be stored: %s", saved_state.url)
         return saved_state.id
 
-    def _add_initial_waterlevel_raster(self, model_id):
-        """Upload initial waterlevel raster and wait for it to be processed."""
-        logger.info("Uploading initial waterlevel raster...")
-        self.waterlevel_raster_id = (
-            self.api.threedimodels_initial_waterlevels_list(model_id).results[0].id
-        )
+    # def add_initial_1d_waterlevels(self, model_id):
+        # logger.info("Uploading initial 1d waterlevel")
+        # ini_wl_api_call = self.api.simulations_initial1d_water_level_predefined_create(
+            # simulation_pk=self.simulation_id, data={}
+        # )
+        # logger.info("Added initial 1D waterlevel %s", ini_wl_api_call.url)
 
-        ini_wl_api_call = self.api.simulations_initial2d_water_level_raster_create(
-            simulation_pk=self.simulation_id,
-            data={
-                "aggregation_method": self.settings.initial_waterlevel,
-                "initial_waterlevel": self.waterlevel_raster_id,
-            },
-        )
-        logger.info("Added initial waterlevel raster to %s", ini_wl_api_call.url)
+    # def _add_initial_waterlevel_raster(self, model_id):
+        # """Upload initial waterlevel raster and wait for it to be processed."""
+        # logger.info("Uploading initial waterlevel raster...")
+        # self.waterlevel_raster_id = (
+            # self.api.threedimodels_initial_waterlevels_list(model_id).results[0].id
+        # )
+
+        # ini_wl_api_call = self.api.simulations_initial2d_water_level_raster_create(
+            # simulation_pk=self.simulation_id,
+            # data={
+                # "aggregation_method": self.settings.initial_waterlevel,
+                # "initial_waterlevel": self.waterlevel_raster_id,
+            # },
+        # )
+        # logger.info("Added initial waterlevel raster to %s", ini_wl_api_call.url)
 
     def _add_netcdf_rain(self, rain_raster_netcdf: Path):
         """Upload rain raster netcdf file and wait for it to be processed."""